third_party/WebKit/LayoutTests/http/tests/inspector/sources/source-code-update-on-frame-reload-expected.txt

CONSOLE MESSAGE: line 5: 0
CONSOLE MESSAGE: line 5: 1
Verifies that UISourceCode gets updated when corresponding frame is reloaded.


Before navigation
====================================
==Source frame contents start==
<html>
<title>iframe</title>
<body>
<script>
var a = '0';console.log(a);
</script>
<div>test</div>
</body>
</html>


==Source frame contents end==
After navigation
====================================
==Source frame contents start==
<html>
<title>iframe</title>
<body>
<script>
var a = '1';console.log(a);
</script>
<div>test</div>
</body>
</html>


==Source frame contents end==
